feat: Technospheres
- add Mnemosphere, Hoplosphere and Mnemosphere Receptacle item types
- add infrastructure for "pseudo documents" nested in items

+++ module/pipelines/effects.mjs
@@ -99,13 +99,23 @@ export async function onManageActiveEffect(event, owner, action) {
 	switch (action ?? actionElement.dataset.action) {
 		case 'create':
 			return createTemporaryEffect(owner, actionElement.dataset.effectType);
-		case 'edit':
-			return resolveEffect().sheet.render(true);
+		case 'edit': {
+			const effect = resolveEffect();
+			return effect.sheet.render(true, { editable: owner === effect.parent });
+		}
 		case 'delete': {
 			const _effect = resolveEffect();
 			if (canBeRemoved(_effect)) {
-				sendToChatEffectRemoved(_effect, owner);
-				return _effect.delete();
+				if (
+					await foundry.applications.api.DialogV2.confirm({
+						title: game.i18n.format('FU.DialogDeleteItemTitle', { item: _effect.name }),
+						content: game.i18n.format('FU.DialogDeleteItemDescription', { item: _effect.name }),
+						rejectClose: false,
+					})
+				) {
+					sendToChatEffectRemoved(_effect, owner);
+					return _effect.delete();
+				}
 			}
 			break;
 		}
@@ -676,4 +686,5 @@ export const Effects = Object.freeze({
 	BOONS_AND_BANES,
 	DAMAGE_TYPES,
 	STATUS_EFFECTS,
+	createTemporaryEffect,
 });
